+++ src/cobib/commands/delete.py
@@ -32,9 +32,6 @@ from __future__ import annotations
 import logging
 import os
 
-from rich.console import Console
-from rich.prompt import PromptBase, PromptType
-from textual.app import App
 from typing_extensions import override
 
 from cobib.config import Event, config
@@ -63,13 +60,8 @@ class DeleteCommand(Command):
     name = "delete"
 
     @override
-    def __init__(
-        self,
-        *args: str,
-        console: Console | App[None] | None = None,
-        prompt: type[PromptBase[PromptType]] | None = None,
-    ) -> None:
-        super().__init__(*args, console=console, prompt=prompt)
+    def __init__(self, *args: str) -> None:
+        super().__init__(*args)
 
         self.deleted_entries: set[str] = set()
         """A set of labels which were deleted by this command."""