+++ src/cobib/commands/redo.py
@@ -26,9 +26,6 @@ import subprocess
 import sys
 from pathlib import Path
 
-from rich.console import Console
-from rich.prompt import PromptBase, PromptType
-from textual.app import App
 from typing_extensions import override
 
 from cobib.config import Event, config
@@ -50,13 +47,8 @@ class RedoCommand(Command):
     name = "redo"
 
     @override
-    def __init__(
-        self,
-        *args: str,
-        console: Console | App[None] | None = None,
-        prompt: type[PromptBase[PromptType]] | None = None,
-    ) -> None:
-        super().__init__(*args, console=console, prompt=prompt)
+    def __init__(self, *args: str) -> None:
+        super().__init__(*args)
 
         self.root: Path
         """The path to the root of the git repository tracking the database."""